RF: move all the tests from reproin.py to its test_
To avoid pytest freaking out, as in
https://travis-ci.org/nipy/heudiconv/jobs/453993426
